requirements.txt
apscheduler
pymysql
flask
coloredlogs
gevent
Flask-Compress
flask_cors